requirements.txt
requests==2.24.0
beautifulsoup4==4.9.1
pymongo==3.11.0
dnspython==2.0.0
lxml==4.5.2
python-dotenv==0.14.0
